apps/chatbot/requirements.txt
langchain==0.1.16
llama-cpp-python==0.2.63
streamlit==1.33.0
huggingface_hub==0.20.3
